Ahora se genera el Treemap
el programa ya puede mostrar el treemap en la GUI

+++ src/project/fileDescriptor/Descriptable.java
@@ -13,7 +13,7 @@ import javafx.scene.paint.Color;
  */
 public interface Descriptable {
     
-    public long getSize();
+    public double getSize();
     public String getDescription();
     public Color getColor();
 }
